The numbers of scheduled meetings and dates for each member are now listed on the main character info screen (commonactions.cpp, commondisplay.cpp, includes.h) Window title is now set to game name and version (game.cpp) MSVC project file updated to compile correctly with newer source files, and compile with STL markup by default so Express (free) versions of MSVC will work (game.vcproj) minor: int casting added to one line to get rid of minor compiler warning (interrogation.cpp)

+++ src/daily/interrogation.cpp
@@ -920,7 +920,7 @@ void tendhostage(Creature *cr,char &clearformess)
          float rapport_temp = rapport[a->id];
 
          if(!techniques[TECHNIQUE_RESTRAIN])attack += 5;
-         attack += rapport[a->id] * 3;
+         attack += int(rapport[a->id] * 3);
 
          y+=1;
          move(y,0);
